pyproject.toml
[project]
name = "genaifundamentals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.46.1",
    "langchain>=0.3.27",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
]
